Align adapter eval docs and CLI
- use JsonDocAdapter in CLI eval path

- store item node handles as usize for adapter refs

- update design doc to describe adapter tradeoffs without perf numbers

+++ src/item.zig
@@ -43,7 +43,7 @@ pub const Item = struct {
     source_end: u32,
     data_pos: u32,
     data_end: u32,
-    node: ?u32,
+    node: ?usize,
     value: ?Value,
 };
